dev-1 Файлы компонетов распределяет по папкам; стилизует страницы; правит логику отображения кнопок
