+++ package.json
@@ -7,7 +7,7 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.27.2",
-    "node-sass": "^7.0.3",
+    "normalize.css": "^8.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.4.1",
@@ -45,9 +45,16 @@
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-react": "^7.31.8",
     "eslint-plugin-react-hooks": "^4.6.0",
+    "node-sass": "^7.0.3",
     "prettier": "2.7.1",
+    "sass-loader": "^13.1.0",
     "stylelint": "^14.13.0",
-    "stylelint-config-htmlacademy": "^1.1.2",
-    "stylelint-config-prettier": "^9.0.3"
+    "stylelint-at-rule-no-children": "^0.3.1",
+    "stylelint-config-prettier": "^9.0.3",
+    "stylelint-config-recommended": "^9.0.0",
+    "stylelint-declaration-block-no-ignored-properties": "^2.5.0",
+    "stylelint-order": "^5.0.0",
+    "stylelint-scss": "^4.3.0",
+    "stylelint-value-no-unknown-custom-properties": "^4.0.0"
   }
 }
