+++ src/components/AuthNav/AuthNav.jsx
@@ -5,27 +5,27 @@ import { useNavigate } from 'react-router-dom';
 import { useAuth } from '../../hooks/useAuth';
 import MyButton from '../UI/button/MyButton.jsx';
 
-
 const AuthNav = () => {
-  const { signOut } = useAuth();
+  const { isAuth, signOut } = useAuth();
   const navigate = useNavigate();
 
   return (
-    <ul className="auth-navigation">
-      <li className="auth-navigation__item">
-        <MyButton className="auth-navigation__link">Login</MyButton>
-      </li>
-      <li className="auth-navigation__item">
+    <div className="auth-navigation">
+      {isAuth ? (
         <MyButton
           onClick={() => {
             signOut(() => navigate('/'));
           }}
-          className="auth-navigation__link"
+          className="auth-navigation__btn"
         >
           Logout
         </MyButton>
-      </li>
-    </ul>
+      ) : (
+        <MyButton className="auth-navigation__btn auth-navigation__btn--green">
+          Login
+        </MyButton>
+      )}
+    </div>
   );
 };
