+++ src/components/AuthNav/AuthNav.scss
@@ -1,8 +1,8 @@
-.auth-navigation {
-	display: flex;
-	align-items: center;
-}
+.auth-navigation__btn {
+	height: 100%;
+	border-radius: 4px;
 
-.auth-navigation__item {
-	margin-right: 12px;
+	&.auth-navigation__btn--green {
+		background-color: $limegreen;
+	}
 }
