+++ src/components/UI/button/MyButton.scss
@@ -1,5 +1,5 @@
 .my-btn {
-	height: 30px;
+	min-height: 30px;
 	background-color: $tomato;
 	border: none;
 	color: $white;
