+++ src/components/UI/navigation/MyNavigation.scss
@@ -1,8 +1,6 @@
 .navigation {
 	display: flex;
-	width: 80%;
-	margin: 0 auto 72px;
-	border-radius: 4px;
+	align-items: stretch;
 }
 
 .navigation__list {
