--- src/pages/About.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const About = () => {
-  return <h1 style={{ textAlign: 'center' }}>About page</h1>;
-};
-
-export default About;
