+++ src/pages/About/About.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const About = () => {
+  return (
+    <div className="about-page page">
+      <div className="container">
+        <h1 className='page__title'>About page</h1>
+        <div>Content ...</div>
+      </div>
+    </div>
+  );
+};
+
+export default About;
