+++ src/pages/About/About.scss
@@ -1,0 +1,5 @@
+.about-page {
+  &__title {
+    margin-bottom: 32px;
+  }
+}
