+++ src/pages/ErrorPage/ErrorPage.scss
@@ -1,0 +1,4 @@
+.error-page {
+	width: 80%;
+	margin: 0 auto;
+}
