+++ src/pages/Layout.jsx
@@ -5,8 +5,11 @@ import MyNavigation from '../components/UI/navigation/MyNavigation.jsx';
 
 const Layout = () => {
   return (
-    <div>
-      <MyNavigation />
+    <div className='layout'>
+      <header className='layout__header'>
+        <MyNavigation />
+      </header>
+      
       <Outlet />
     </div>
   );
