--- src/pages/Main.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Main = () => {
-  return <h1 style={{ textAlign: 'center' }}>Main page</h1>;
-};
-
-export default Main;
