+++ src/pages/Main/Main.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+const Main = () => {
+  return (
+    <div className="main-page page">
+      <div className="container">
+        <h1 className='main-page__title page__title'>Main page</h1>
+        <div className='main-page__content'>Content ...</div>
+      </div>
+    </div>
+  );
+};
+
+export default Main;
