+++ src/pages/Main/Main.scss
@@ -1,0 +1,5 @@
+.main-page {
+  &__title {
+    margin-bottom: 32px;
+  }
+}
