+++ src/styles/App.scss
@@ -1,4 +1,16 @@
-@import './variables.scss';
+@import './variables/colors.scss';
+@import './variables/mixins.scss';
+@import './variables/media-query.scss';
+@import './typografy.scss';
+@import './animations/spinnerAnimation.scss';
+
+@import './layout.scss';
+@import '../pages/Posts/Posts.scss';
+@import '../pages/ErrorPage/ErrorPage.scss';
+@import '../pages/Login/Login.scss';
+@import '../pages/Main/Main.scss';
+@import '../pages/About/About.scss';
+
 @import '../components/AuthNav/AuthNav.scss';
 @import '../components/PostList/PostList.scss';
 @import '../components/PostItem/PostItem.scss';
@@ -13,65 +25,41 @@
 @import '../components/UI/pagination/MyPagination.scss';
 @import '../components/UI/select/MySelect.scss';
 
+html,
 body {
-	margin: 0;
-	padding: 0;
+  font-size: 16px;
+  line-height: 16px;
+  background-color: $white;
+  margin: 0;
+  padding: 0;
+}
+
+body {
+  position: relative;
+  font-weight: 400;
 }
 
 * {
-	box-sizing: border-box;
+  box-sizing: border-box;
 }
 
 ul {
-	list-style: none;
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
+
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+  margin: 0;
+  padding: 0;
+}
+
+p {
 	padding: 0;
 	margin: 0;
 }
-
-//  Todo replace to container
-#root {
-	display: flex;
-	flex-direction: column;
-	padding-top: 20px;
-}
-
-.posts-page,
-.error-page {
-	width: 80%;
-	margin: 0 auto;
-	padding: 0 0 60px;
-	position: relative;
-}
-
-.posts-page__btns {
-	display: flex;
-	margin-bottom: 30px;
-}
-
-.posts-page__btn {
-	margin-right: 8px;
-	border-radius: 4px;
-}
-
-.posts-page__btn--create {
-	background-color: $limegreen;
-}
-
-.loader-wrapper {
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	align-items: center;
-}
-
-/* login */
-
-.form-login {
-	margin-bottom: 30px;
-}
-
-.form-login__btn {
-	padding: 4px 12px;
-	border-radius: 4px;
-}
-/* login end */
