+++ src/styles/animations/spinnerAnimation.scss
@@ -1,0 +1,9 @@
+@keyframes lds-dual-ring {
+  0% {
+    transform: rotate(0deg);
+  }
+
+  100% {
+    transform: rotate(360deg);
+  }
+}
