+++ src/styles/typografy.scss
@@ -1,0 +1,11 @@
+.page {
+  &__title {
+    margin-bottom: 32px;
+    @include font(24, 32);
+
+    @media #{$lg-and-up} {
+      margin-bottom: 54px;
+      @include font(36, 42);
+    }
+  }
+}
